api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "ArthurSousa460",email = "arthurprogrammer460@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
